+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/frostserver/service/PluginManager.java
@@ -142,7 +142,9 @@ public class PluginManager implements ConfigDefaults {
         LOGGER.info("Loading plugins...");
         loadPlugins(provided);
         loadPlugins(extra);
-        initPlugins(PersistenceManagerFactory.getInstance(settings).create());
+        try (PersistenceManager pm = PersistenceManagerFactory.getInstance(settings).create()) {
+            initPlugins(pm);
+        }
     }
 
     public void initPlugins(PersistenceManager pm) {